added last commands for custom voice

+++ src/event_listeners/onVoiceStateUpdate.js
@@ -28,6 +28,8 @@ module.exports = (bot) => {
                 // Check if memeber's new voice channel is equal to one of the active channels
                 if (guildSettings.channels.includes(channelID)) {
                     const defaultName = `${newState.member.user.username}'s Channel`;
+                    const joinMsg = "**%user% joined!**";
+                    const leaveMsg = "**%user% left!**";
                     let userSettings = bot.dbClient.getUserSettings(
                         newState.id
                     );
@@ -90,6 +92,8 @@ module.exports = (bot) => {
                             owner: newState.member.user.id,
                             VCName: userVoiceChannel.name,
                             TCName: userTextChannel.name,
+                            joinMsg,
+                            leaveMsg,
                             userLimit: 0,
                             bitrate: 64000,
                         };
@@ -110,16 +114,42 @@ module.exports = (bot) => {
                             Utilities.channelInfo(newState.guild, userSettings)
                         );
                     } else {
+                        // The user has data in the db
                         // Voice channel creation
                         const userVoiceChannel = await newState.guild.channels.create(
                             userSettings.VCName,
                             {
                                 type: "voice",
                                 parent: guildSettings.categoryID,
-                                bitrate: userSettings.bitrate,
                                 userLimit: userSettings.userLimit,
                             }
                         );
+
+                        let MAX_BITRATE;
+                        if (newState.guild.premiumTier === 0) {
+                            MAX_BITRATE = 96;
+                        } else if (newState.guild.premiumTier === 1) {
+                            MAX_BITRATE = 128;
+                        } else if (newState.guild.premiumTier === 2) {
+                            MAX_BITRATE = 256;
+                        } else if (newState.guild.premiumTier === 3) {
+                            MAX_BITRATE = 384;
+                        }
+
+                        let warningEmbed;
+                        if (userSettings.bitrate > MAX_BITRATE * 1000) {
+                            userVoiceChannel.setBitrate(MAX_BITRATE * 1000);
+                            warningEmbed = new MessageEmbed()
+                                .setTitle("Warning!")
+                                .setColor(colors.Yellow)
+                                .setDescription(
+                                    stripIndents`⚠ **Your bitrate is too powerful for this server to handle! I've reduced it to \`${MAX_BITRATE}kbps!\`**
+        
+                                    *NOTE: Your original bitrate is not changed.*`
+                                );
+                        } else {
+                            userVoiceChannel.setBitrate(userSettings.bitrate);
+                        }
 
                         // Text channel creation
                         const userTextChannel = await newState.guild.channels.create(
@@ -175,13 +205,19 @@ module.exports = (bot) => {
                         userTextChannel.send(
                             Utilities.channelInfo(newState.guild, userSettings)
                         );
+                        if (warningEmbed) {
+                            userTextChannel.send(warningEmbed);
+                        }
                     }
                 } else if (activeChannels) {
-                    if (
-                        activeChannels.find(
-                            (channel) => channel.voice === channelID
-                        )
-                    ) {
+                    const currentChannel = activeChannels.find(
+                        (channel) => channel.voice === channelID
+                    );
+                    if (currentChannel) {
+                        const userSettings = bot.dbClient.getUserSettings(
+                            currentChannel.owner
+                        );
+
                         if (newState.channel.members.size > 1) {
                             const textChannel = oldState.guild.channels.resolve(
                                 activeChannels.find(
@@ -196,7 +232,10 @@ module.exports = (bot) => {
                             const embed = new MessageEmbed()
                                 .setColor(colors.Turquoise)
                                 .setDescription(
-                                    stripIndents`**${newState.member} joined!**`
+                                    Utilities.replacePlaceholders(
+                                        newState,
+                                        userSettings.joinMsg
+                                    )
                                 );
                             textChannel.send(embed);
                         }
@@ -210,11 +249,16 @@ module.exports = (bot) => {
             const channelID = oldState.channelID;
             const guildID = oldState.guild.id;
 
-            // Fetch guild settings and active channels from local db
-
+            // Fetch user settings and active channels from local db
             const activeChannels = bot.dbClient.getActiveChannels(guildID);
+            const currentChannel = activeChannels.find(
+                (channel) => channel.voice === channelID
+            );
 
-            if (activeChannels.find((channel) => channel.voice === channelID)) {
+            if (currentChannel) {
+                const userSettings = bot.dbClient.getUserSettings(
+                    currentChannel.owner
+                );
                 const voiceChannel = oldState.guild.channels.resolve(channelID);
                 const textChannelID = activeChannels.find(
                     (channel) => channel.voice === channelID
@@ -237,11 +281,16 @@ module.exports = (bot) => {
                         (permission) =>
                             permission.id === oldState.member.user.id
                     );
-                    userPerms.delete();
+                    if (userPerms) {
+                        userPerms.delete();
+                    }
                     const embed = new MessageEmbed()
                         .setColor(colors.Turquoise)
                         .setDescription(
-                            stripIndents`**${oldState.member} left!**`
+                            Utilities.replacePlaceholders(
+                                oldState,
+                                userSettings.leaveMsg
+                            )
                         );
                     textChannel.send(embed);
                 }
